requirements.txt
matplotlib
numpy
pytorch
torchvision
typer
tqdm